+++ spec/type_spec.rb
@@ -1,6 +1,23 @@
 # frozen_string_literal: true
 
 RSpec.describe KSUID::Type do
+  describe 'value object semantics' do
+    it 'uses value comparison instead of identity comparison' do
+      ksuid1 = KSUID.new(time: Time.now)
+      ksuid2 = KSUID.from_base62(ksuid1.to_s)
+      hash = {}
+
+      hash[ksuid1] = 'Hello, world'
+
+      aggregate_failures do
+        expect(ksuid1).to eq ksuid2
+        expect(ksuid1).to eql ksuid2
+        expect(ksuid1).not_to equal ksuid2
+        expect(hash[ksuid2]).to eq 'Hello, world'
+      end
+    end
+  end
+
   describe '.from_base62' do
     it 'converts a base62 KSUID properly' do
       ksuid = KSUID.from_base62(KSUID::MAX_STRING_ENCODED)
@@ -17,6 +34,18 @@ RSpec.describe KSUID::Type do
       array = [ksuid2, ksuid1].sort
 
       expect(array).to eq([ksuid1, ksuid2])
+    end
+  end
+
+  describe '#==' do
+    it 'matches against other KSUID::Types as well as String' do
+      ksuid1 = KSUID.new(time: Time.now)
+      ksuid2 = KSUID.from_base62(ksuid1.to_s)
+
+      aggregate_failures do
+        expect(ksuid1).to eq ksuid2
+        expect(ksuid1).to eq ksuid2.to_s
+      end
     end
   end